+++ src/lies/builders/__init__.py
@@ -11,6 +11,7 @@ from __future__ import annotations
 
 from . import bespoke as _bespoke  # noqa: F401
 from . import html as _html  # noqa: F401
+from . import liquid as _liquid  # noqa: F401
 from . import pdf as _pdf  # noqa: F401
 from . import sphinx as _sphinx  # noqa: F401